Add AWS Kinesis to WbFlux SSE sample
https://stackoverflow.com/questions/51669324/kinesis-as-producer-in-spring-boot-reactive-stream-api

This sample demonstrate a simple bridging of AWS Kinesis stream records
to the Server Side Events subscribers.
The `@StreamListener` sink side is based on the Spring Cloud Stream
Reactive support streaming incoming message to the `Flux` argument.

Since the `Flux` returned by the `MessageChannelToInputFluxParameterAdapter`
is a `FluxPublish`, so it is safe to let this one be subscribed as many
times as we need and all the subscribers will get the same data on the
fly, which is actually an SSE purpose.

The test-case is based on the locally ran Kinesalite and also adds
extra Source binding for producing message into the AWS Kinesis stream.
The verification is done by the auto-configured `WebTestClient`